Merge from TaikoLocalServer-CHN repo

+++ TaikoLocalServer/Services/Interfaces/IAiDatumService.cs
@@ -4,9 +4,9 @@ namespace TaikoLocalServer.Services.Interfaces;
 
 public interface IAiDatumService
 {
-    public Task<List<AiScoreDatum>> GetAllAiScoreById(uint baid);
+    public Task<List<AiScoreDatum>> GetAllAiScoreById(ulong baid);
 
-    public Task<AiScoreDatum?> GetSongAiScore(uint baid, uint songId, Difficulty difficulty);
+    public Task<AiScoreDatum?> GetSongAiScore(ulong baid, uint songId, Difficulty difficulty);
 
     public Task UpdateSongAiScore(AiScoreDatum datum);
