+++ TaikoLocalServer/Services/Interfaces/ISongBestDatumService.cs
@@ -5,9 +5,9 @@ namespace TaikoLocalServer.Services.Interfaces;
 
 public interface ISongBestDatumService
 {
-    public Task<List<SongBestDatum>> GetAllSongBestData(uint baid);
+    public Task<List<SongBestDatum>> GetAllSongBestData(ulong baid);
 
     public Task UpdateOrInsertSongBestDatum(SongBestDatum datum);
 
-    public Task<List<SongBestData>> GetAllSongBestAsModel(uint baid);
+    public Task<List<SongBestData>> GetAllSongBestAsModel(ulong baid);
 }
